requirements.txt
django-adminplus==0.5